Auto cleaning of configuration tables (#49)
When dropping schema, role or tablespace, the corresponding entries from
diskquota.quota_config and diskquota.target were not deleted. The fact that they
remain may lead to an unexpected quota on a new scheme, role or tablespace with
the same oid (in the case of oid wraparound).

Add cleaning of configuration tables when invalid record is detected during data
loading from these tables.

Cleaning at the loading time only is not enough, because scheme, role or
tablespace can be dropped and object of the same type with the same oid can be
created in the period between two loadings. It is possible to use
object_access_hook to handle dropping, but it is difficult, this case is very
unlikely and unexpected quota is not critical problem, because it can be removed
by DBA. So the patch doesn't handle this case.

Ticket: ADBDEV-7208